% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_finances}
\alias{load_finances}
\title{load_finances}
\usage{
load_finances(finances, file_path = NULL)
}
\arguments{
\item{finances}{named data_list from `finances$save_excel`. If NULL will load
blank. If missing (default) it will load sample. Otherwise will use provided
data.}

\item{file_path}{Optional. if file_path provided will load from that file. If
provided, will ignore any finances object provided to function}
}
\description{
load_finances
}
