% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RosyFinance-package.R
\docType{package}
\name{RosyFinance-package}
\alias{RosyFinance}
\alias{RosyFinance-package}
\title{RosyFinance: An Amazing Shiny App}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Brandon Rose \email{thecodingdocs@gmail.com} (\href{https://orcid.org/0009-0009-7813-1960}{ORCID})

Authors:
\itemize{
  \item Brandon Rose \email{thecodingdocs@gmail.com} (\href{https://orcid.org/0009-0009-7813-1960}{ORCID})
}

}
\keyword{internal}
